Separate rewrites: scalar linalg folding, scalar tensor folding, cmpi extui folding
